# Report aggregate statistics for solution as well as some solution perf numbers